requirements.txt
Django==3.2.9
djangorestframework==3.12.4
mysqlclient==2.1.0
icecream==2.1.1
googlemaps==4.5.3
pandas
django-cors-headers